Title: 길이 단위 환산하기, Time: 45ms, Memory: 16MB, Status: Passed - Codetree

+++ 250422/길이 단위 환산하기/convert-length-units.py
@@ -1,3 +1,3 @@
-ft = double(input())
+ft = float(input())
 
 print(f'{ft * 30.48:.1f}')
